# web/openapi.json
{
  "components": {
    "schemas": {
...
      "BacktestActualExit": {
        "additionalProperties": false,
        "properties": {
          "exit_ev": {
            "minimum": 0.0,
            "title": "Exit Ev",
            "type": "number"
          },
          "irr": {
            "description": "Actual IRR (%)",
            "title": "Irr",
            "type": "number"
          },
          "moic": {
            "minimum": 0.0,
            "title": "Moic",
            "type": "number"
          },
          "net_debt_at_exit": {
            "minimum": 0.0,
            "title": "Net Debt At Exit",
            "type": "number"
          },
          "sponsor_equity_entry": {
            "minimum": 0.0,
            "title": "Sponsor Equity Entry",
            "type": "number"
          }
        },
        "required": [
          "exit_ev",
          "net_debt_at_exit",
          "sponsor_equity_entry",
          "moic",
          "irr"
        ],
        "title": "BacktestActualExit",
        "type": "object"
      },
      "BacktestActuals": {
        "additionalProperties": false,
        "properties": {
          "ebitda": {
            "items": {
              "type": "number"
            },
            "title": "Ebitda",
            "type": "array"
          },
          "fcf": {
            "items": {
              "type": "number"
            },
            "title": "Fcf",
            "type": "array"
          },
          "net_income": {
            "items": {
              "type": "number"
            },
            "title": "Net Income",
            "type": "array"
          },
          "revenue": {
            "items": {
              "type": "number"
            },
            "title": "Revenue",
            "type": "array"
          },
          "total_debt": {
            "items": {
              "type": "number"
            },
            "title": "Total Debt",
            "type": "array"
          }
        },
        "required": [
          "revenue",
          "ebitda",
          "net_income",
          "fcf",
          "total_debt"
        ],
        "title": "BacktestActuals",
        "type": "object"
      },
      "BacktestEntry": {
        "additionalProperties": false,
        "properties": {
          "base_rate": {
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Base Rate",
            "type": "number"
          },
          "capex_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Capex Pct",
            "type": "number"
          },
          "da_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Da Pct",
            "type": "number"
          },
          "debt_pct": {
            "maximum": 99.0,
            "minimum": 0.0,
            "title": "Debt Pct",
            "type": "number"
          },
          "entry_ebitda": {
            "exclusiveMinimum": 0.0,
            "title": "Entry Ebitda",
            "type": "number"
          },
          "entry_multiple": {
            "exclusiveMinimum": 0.0,
            "title": "Entry Multiple",
            "type": "number"
          },
          "exit_multiple": {
            "exclusiveMinimum": 0.0,
            "title": "Exit Multiple",
            "type": "number"
          },
          "gross_margin": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Gross Margin",
            "type": "number"
          },
          "holding_period": {
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Holding Period",
            "type": "integer"
          },
          "mezz_spread": {
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Mezz Spread",
            "type": "number"
          },
          "nwc_pct": {
            "title": "Nwc Pct",
            "type": "number"
          },
          "opex_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Opex Pct",
            "type": "number"
          },
          "revenue_growth": {
            "title": "Revenue Growth",
            "type": "number"
          },
          "senior_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Senior Pct",
            "type": "number"
          },
          "tax_rate": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Tax Rate",
            "type": "number"
          }
        },
        "required": [
          "entry_ebitda",
          "entry_multiple",
          "exit_multiple",
          "holding_period",
          "debt_pct",
          "senior_pct",
          "base_rate",
          "mezz_spread",
          "revenue_growth",
          "gross_margin",
          "opex_pct",
          "da_pct",
          "tax_rate",
          "capex_pct",
          "nwc_pct"
        ],
        "title": "BacktestEntry",
        "type": "object"
      },
      "BacktestRequest": {
        "additionalProperties": false,
        "properties": {
          "actual": {
            "$ref": "#/components/schemas/BacktestActuals"
          },
          "actual_exit": {
            "$ref": "#/components/schemas/BacktestActualExit"
          },
          "entry": {
            "$ref": "#/components/schemas/BacktestEntry"
          },
          "histogram_bins": {
            "default": 80,
            "maximum": 400.0,
            "minimum": 10.0,
            "title": "Histogram Bins",
            "type": "integer"
          },
          "n": {
            "default": 30000,
            "maximum": 100000.0,
            "minimum": 1000.0,
            "title": "N",
            "type": "integer"
          },
          "settings": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "default": {},
            "title": "Settings",
            "type": "object"
          }
        },
        "required": [
          "entry",
          "actual",
          "actual_exit"
        ],
        "title": "BacktestRequest",
        "type": "object"
      },
      "BacktestResponse": {
        "properties": {
          "actual_ebitda_margin": {
            "items": {
              "type": "number"
            },
            "title": "Actual Ebitda Margin",
            "type": "array"
          },
          "actual_equity_entry": {
            "title": "Actual Equity Entry",
            "type": "number"
          },
          "actual_exit_equity": {
            "title": "Actual Exit Equity",
            "type": "number"
          },
          "actual_exit_multiple": {
            "title": "Actual Exit Multiple",
            "type": "number"
          },
          "actual_irr": {
            "title": "Actual Irr",
            "type": "number"
          },
          "actual_moic": {
            "title": "Actual Moic",
            "type": "number"
          },
          "actual_percentile": {
            "title": "Actual Percentile",
            "type": "number"
          },
          "attribution": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Exact split of actual minus predicted exit equity ($M): exit_ebitda, exit_multiple, net_debt",
            "title": "Attribution",
            "type": "object"
          },
          "irr_histogram": {
            "$ref": "#/components/schemas/Histogram"
          },
          "predicted_ebitda": {
            "items": {
              "type": "number"
            },
            "title": "Predicted Ebitda",
            "type": "array"
          },
          "predicted_ebitda_margin": {
            "title": "Predicted Ebitda Margin",
            "type": "number"
          },
          "predicted_equity_entry": {
            "title": "Predicted Equity Entry",
            "type": "number"
          },
          "predicted_exit_equity": {
            "title": "Predicted Exit Equity",
            "type": "number"
          },
          "predicted_irr_mean": {
            "title": "Predicted Irr Mean",
            "type": "number"
          },
          "predicted_irr_p5": {
            "title": "Predicted Irr P5",
            "type": "number"
          },
          "predicted_irr_p95": {
            "title": "Predicted Irr P95",
            "type": "number"
          },
          "predicted_moic": {
            "title": "Predicted Moic",
            "type": "number"
          },
          "predicted_net_debt_at_exit": {
            "title": "Predicted Net Debt At Exit",
            "type": "number"
          },
          "years": {
            "items": {
              "$ref": "#/components/schemas/BacktestYear"
            },
            "title": "Years",
            "type": "array"
          }
        },
        "required": [
          "predicted_irr_mean",
          "predicted_irr_p5",
          "predicted_irr_p95",
          "predicted_moic",
          "predicted_equity_entry",
          "predicted_exit_equity",
          "predicted_ebitda",
          "actual_irr",
          "actual_moic",
          "actual_equity_entry",
          "actual_exit_equity",
          "actual_percentile",
          "actual_ebitda_margin",
          "predicted_ebitda_margin",
          "predicted_net_debt_at_exit",
          "actual_exit_multiple",
          "attribution",
          "irr_histogram",
          "years"
        ],
        "title": "BacktestResponse",
        "type": "object"
      },
      "BacktestYear": {
        "properties": {
          "actual_ebitda": {
            "title": "Actual Ebitda",
            "type": "number"
          },
          "actual_fcf": {
            "title": "Actual Fcf",
            "type": "number"
          },
          "actual_revenue": {
            "title": "Actual Revenue",
            "type": "number"
          },
          "actual_total_debt": {
            "title": "Actual Total Debt",
            "type": "number"
          },
          "ebitda_variance": {
            "title": "Ebitda Variance",
            "type": "number"
          },
          "predicted_ebitda": {
            "title": "Predicted Ebitda",
            "type": "number"
          },
          "year_index": {
            "title": "Year Index",
            "type": "integer"
          }
        },
        "required": [
          "year_index",
          "predicted_ebitda",
          "actual_ebitda",
          "ebitda_variance",
          "actual_revenue",
          "actual_fcf",
          "actual_total_debt"
        ],
        "title": "BacktestYear",
        "type": "object"
      },
...
      "DealCreate": {
        "additionalProperties": false,
        "properties": {
          "inputs": {
            "$ref": "#/components/schemas/DealInputsIn"
          },
          "name": {
            "description": "Shown in the deal list",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "settings": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "default": {},
            "title": "Settings",
            "type": "object"
          }
        },
        "required": [
          "inputs",
          "name"
        ],
        "title": "DealCreate",
        "type": "object"
      },
      "DealDetail": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "created_at": {
            "description": "UTC, ISO 8601",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inputs": {
            "$ref": "#/components/schemas/DealInputsIn"
          },
          "latest_version": {
            "description": "Number of the newest version",
            "title": "Latest Version",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "settings": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Settings",
            "type": "object"
          },
          "updated_at": {
            "description": "UTC, ISO 8601",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "archived",
          "latest_version",
          "created_at",
          "updated_at",
          "inputs",
          "settings"
        ],
        "title": "DealDetail",
        "type": "object"
      },
...
      "DealInputsIn": {
        "additionalProperties": false,
        "description": "Deal wizard inputs.\n\ndebt_pct and senior_pct default to the model's stored defaults (60% / 70%).\nThe Streamlit wizard instead derives them on load from its financing\ninputs -- 3.4x senior + 0.8x mezz EBITDA, i.e. 42% and ~81% -- which is why\nits default deal shows a lower IRR. Use /deal/sources-and-uses to derive\nthem from debt multiples the same way.",
        "properties": {
          "ap_days": {
            "default": 60.0,
            "maximum": 365.0,
            "minimum": 0.0,
            "title": "Ap Days",
            "type": "number"
          },
          "ar_days": {
            "default": 45.0,
            "maximum": 365.0,
            "minimum": 0.0,
            "title": "Ar Days",
            "type": "number"
          },
          "base_rate": {
            "default": 6.5,
            "description": "Senior interest rate (%)",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Base Rate",
            "type": "number"
          },
          "capex": {
            "default": 4.0,
            "description": "Capex / revenue (%)",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Capex",
            "type": "number"
          },
          "da": {
            "default": 4.0,
            "description": "D&A / revenue (%)",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Da",
            "type": "number"
          },
          "debt_pct": {
            "default": 60.0,
            "description": "Total debt / EV (%)",
            "maximum": 99.0,
            "minimum": 0.0,
            "title": "Debt Pct",
            "type": "number"
          },
          "ebitda": {
            "default": 100.0,
            "description": "LTM EBITDA ($M)",
            "exclusiveMinimum": 0.0,
            "title": "Ebitda",
            "type": "number"
          },
          "entry_mult": {
            "default": 10.0,
            "description": "Entry EV / EBITDA (x)",
            "exclusiveMinimum": 0.0,
            "title": "Entry Mult",
            "type": "number"
          },
          "exit_mult": {
            "default": 11.0,
            "description": "Exit EV / EBITDA (x)",
            "exclusiveMinimum": 0.0,
            "title": "Exit Mult",
            "type": "number"
          },
          "gross_margin": {
            "default": 40.0,
            "description": "Gross margin (%)",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Gross Margin",
            "type": "number"
          },
          "growth": {
            "default": 5.0,
            "description": "Revenue growth (%)",
            "maximum": 100.0,
            "minimum": -50.0,
            "title": "Growth",
            "type": "number"
          },
          "hold": {
            "default": 5,
            "description": "Holding period (years)",
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Hold",
            "type": "integer"
          },
          "inv_days": {
            "default": 30.0,
            "maximum": 365.0,
            "minimum": 0.0,
            "title": "Inv Days",
            "type": "number"
          },
          "mezz_spread": {
            "default": 4.0,
            "description": "Mezz spread over senior (%)",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Mezz Spread",
            "type": "number"
          },
          "mincash": {
            "default": 0.0,
            "description": "Minimum cash ($M)",
            "minimum": 0.0,
            "title": "Mincash",
            "type": "number"
          },
          "nwc": {
            "default": 1.0,
            "description": "Change in NWC / revenue (%)",
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Nwc",
            "type": "number"
          },
          "opex": {
            "default": 18.0,
            "description": "OpEx / revenue (%)",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Opex",
            "type": "number"
          },
          "senior_pct": {
            "default": 70.0,
            "description": "Senior / total debt (%)",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Senior Pct",
            "type": "number"
          },
          "tax": {
            "default": 25.0,
            "description": "Tax rate (%)",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Tax",
            "type": "number"
          },
          "wsp_mode": {
            "default": false,
            "description": "Use AR/inventory/AP days instead of flat NWC",
            "title": "Wsp Mode",
            "type": "boolean"
          }
        },
        "title": "DealInputsIn",
        "type": "object"
      },
      "DealList": {
        "properties": {
          "deals": {
            "items": {
              "$ref": "#/components/schemas/DealSummary"
            },
            "title": "Deals",
            "type": "array"
          }
        },
        "required": [
          "deals"
        ],
        "title": "DealList",
        "type": "object"
      },
...
      "DealSummary": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "created_at": {
            "description": "UTC, ISO 8601",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_version": {
            "description": "Number of the newest version",
            "title": "Latest Version",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "description": "UTC, ISO 8601",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "archived",
          "latest_version",
          "created_at",
          "updated_at"
        ],
        "title": "DealSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Histogram": {
        "properties": {
          "density": {
            "items": {
              "type": "number"
            },
            "title": "Density",
            "type": "array"
          },
          "edges": {
            "items": {
              "type": "number"
            },
            "title": "Edges",
            "type": "array"
          }
        },
        "required": [
          "edges",
          "density"
        ],
        "title": "Histogram",
        "type": "object"
      },
...
      "LimitInfo": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "retry_after_s": {
            "title": "Retry After S",
            "type": "integer"
          },
          "rule": {
            "title": "Rule",
            "type": "string"
          },
          "window_s": {
            "title": "Window S",
            "type": "integer"
          }
        },
        "required": [
          "rule",
          "limit",
          "window_s",
          "retry_after_s"
        ],
        "title": "LimitInfo",
        "type": "object"
      },
      "LimitRefusal": {
        "description": "Body of a 429 response.",
        "properties": {
          "detail": {
            "description": "What was refused and when to try again, for people",
            "title": "Detail",
            "type": "string"
          },
          "limit": {
            "$ref": "#/components/schemas/LimitInfo"
          }
        },
        "required": [
          "detail",
          "limit"
        ],
        "title": "LimitRefusal",
        "type": "object"
      },
...
      "PreloadedDeal": {
        "properties": {
          "actual": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "default": {},
            "title": "Actual",
            "type": "object"
          },
          "actual_exit": {
            "additionalProperties": {
              "type": "number"
            },
            "default": {},
            "title": "Actual Exit",
            "type": "object"
          },
          "actual_years": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Actual Years",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "entry": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Entry",
            "type": "object"
          },
          "geography": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geography"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          }
        },
        "required": [
          "name",
          "entry"
        ],
        "title": "PreloadedDeal",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Clerk session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "LBO modelling: deal wizard, Monte Carlo, backtesting and 3-statement forecasting.",
    "title": "Simulation Model API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/backtesting/deals": {
      "get": {
        "description": "Historical deals with their entry assumptions and actual results.",
        "operationId": "get_deals_api_backtesting_deals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PreloadedDeal"
                  },
                  "title": "Response Get Deals Api Backtesting Deals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitRefusal"
                }
              }
            },
            "description": "A usage limit was reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Deals",
        "tags": [
          "backtesting"
        ]
      }
    },
    "/api/backtesting/run": {
      "post": {
        "description": "Predict a deal from its entry assumptions and compare with what happened.",
        "operationId": "post_run_api_backtesting_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitRefusal"
                }
              }
            },
            "description": "A usage limit was reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Post Run",
        "tags": [
          "backtesting"
        ]
      }
    },
...
    "/api/deals": {
      "get": {
        "description": "The caller's deals, most recently edited first. ``archived=true`` includes archived ones.",
        "operationId": "list_deals_api_deals_get",
        "parameters": [
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitRefusal"
                }
              }
            },
            "description": "A usage limit was reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Deals",
        "tags": [
          "saved deals"
        ]
      },
      "post": {
        "description": "Save a new deal; its first version is created with it.",
        "operationId": "create_deal_api_deals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DealCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitRefusal"
                }
              }
            },
            "description": "A usage limit was reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Deal",
        "tags": [
          "saved deals"
        ]
      }
    },
...
  }
}